requirementswin.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.7.0
anyio==4.3.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.1.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.11.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
bottleneck==1.3.7
brotli==1.0.9
bson==0.5.10
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
cryptography==42.0.5
dataclasses-json==0.6.6
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.111.0
fastapi-cli==0.0.4
fastembed==0.2.6
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.7
gitpython==3.1.37
google-api-core==2.18.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.56.4
graphlib-backport==1.1.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub>=0.20,<0.21
humanfriendly==10.0
idna==3.7
importlib-metadata==6.8.0
importlib_resources==6.1.1
ipykernel==6.25.1
ipython==8.14.0
ipython_genutils==0.2.0
ipywidgets==7.8.1
jedi==0.19.0
jinja2==3.1.4
joblib==1.4.2
jpype1==1.5.0
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.0
jupyter_events==0.8.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab_pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab_widgets==1.0.0
konlpy==0.6.0
kubernetes==29.0.0
langchain==0.2.0
langchain-anthropic==0.1.13
langchain-community==0.2.0
langchain-core>=0.2.0,<0.3.0
langchain-experimental
langchain-google-community>=1.0.3
langchain-openai==0.1.7
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langsmith==0.1.60
lark==1.1.9
loguru==0.7.2
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.21.2
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mmh3==4.1.0
motor==3.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.8
notebook-shim==0.2.3
numexpr==2.8.4
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.17.3
openai==1.30.1
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==3.2.2
pip==23.3.1
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
posthog==3.5.0
prometheus_client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.7.1
pydantic-core==2.18.2
pydeck==0.8.0
pygments==2.16.1
pymongo==4.7.2
pymupdf==1.24.2
pymupdfb==1.24.1
pymysql==1.1.0
pyopenssl==24.0.0
pyparsing==3.1.2
pypdf==4.2.0
pypika==0.48.9
pyproject-hooks==1.1.0
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pyu2f==0.1.5
pyyaml==6.0.1
rank-bm25==0.2.2
referencing==0.30.2
requests==2.32.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
rsa==4.7.2
safetensors==0.4.3
scikit-learn==1.3.2
selenium==4.20.0
send2trash==1.8.2
sentence-transformers==2.7.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==4.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlalchemy==2.0.30
stack-data==0.6.2
starlette==0.37.2
streamlit==1.32.0
sympy==1.12
tavily-python==0.3.3
tenacity==8.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers>=0.13.2,<0.16.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.3.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.10.0
transformers>=4.34.0,<5.0.0
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing-extensions>=4.9.0
typing_inspect==0.9.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.6
websockets==12.0
wheel==0.43.0
widgetsnbextension==3.6.6
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0







